Menu is now filled from the database. Images are currently all the same until I have access to correct permissions for images.
I adjusted menuItem class to use a decimal type for the price instead of int. The interface will have to be adjusted to reflect this correctly.

+++ RestaurantManagement/OMS-Client/menuItem.cs
@@ -10,7 +10,7 @@ namespace OMS
 		public string name { get; set; }
 		public string description { get; set; }
 		public ImageSource imgSource { get; set; }
-		public int price { get; set; }
+		public decimal price { get; set; }          // It's very important that any number you store here must have an 'm' after it like so price = 69.69m  <------
 		public string category { get; set; }
 		public bool visible { get; set; }
 	}
